Fix local hot reload rust-analyzer issue
Rust-analyzer plugin on vscode runs cargo check on each file in the dir
on save.
This causes the next run of cargo run to have to rebuild every file in
the docker image (something to do with different environments [local and
docker] invalidating checks).
Now there is a separate build dir for local checks (for rust-analyzer)
and for docker's actual build.