Huge rewrite / optimisation which allows response mutation through middleware
Previously, copies of requests and responses caused unnecessary
overhead. Realistically we only needed one mutatable copy of each.
We want to be able to mutate the response in each middleware that does
not return the response.

Async functions / futures could technically (if not awaited) outlive
a tokio task execution, so needs to be hidden behind an Arc and copied
into the task.
Async functions behind an Arc could be called whenever, so any params
passed to it must also be Arc, and available whenever (and also in thise
case Mutex, since we intend to change response).
Modified macros also to still make addition of a Route / middleware just
as easy. We now get to mutate a request or response in each middleware
if we wish.
To return the response, we use response.send();
After each middleware / route, we will check if this has been called and
if so return the response object.